fix(components): rename tooltip component
