+++ src/lib/components/button/AdminButton.tsx
@@ -3,7 +3,7 @@ import { useWallet } from "@cosmos-kit/react";
 
 import { CustomIcon } from "../icon";
 import { ClearAdminModal } from "../modal/contract/ClearAdmin";
-import { Tooltip } from "../TooltipComponent";
+import { Tooltip } from "../Tooltip";
 import { useInternalNavigate } from "lib/app-provider";
 import type { Addr, ContractAddr, Option } from "lib/types";
